requirements.txt
twitchAPI
discord.py
PyYAML
psycopg2